requirements.txt
google-api-python-client==2.114.0
google-auth==2.28.2
requests==2.31.0
